Added more flexible api to upgradeteamsbanner component
