+++ packages/features/teams/index.ts
@@ -1,0 +1,3 @@
+export * from "./components/UpgradeToTeamsBanner";
+export * from "./components/TeamFeatures";
+export * from "./hooks/useIsInTeam";
